+++ rc/build.h
@@ -3,5 +3,5 @@
 #define STRINGIZE(s) STRINGIZE2(s)
 #define VERSION_MAJOR            5
 #define VERSION_MINOR            7
-#define VERSION_REVISION         5
-#define VERSION_BUILD            2178
+#define VERSION_REVISION         6
+#define VERSION_BUILD            2198